+++ R/SIG.conversion.l932dec.R
@@ -1,7 +1,7 @@
 #' Conversion de coordonnées en Lambert 93 vers degrés décimaux
 #'
 #' Cette fonction permet de convertir des coordonnées en du Lambert 93 (2154) vers degrés décimaux (4326)
-#' @name SIG.conversion.l93dec2
+#' @name SIG.conversion.l932dec
 #' @param coord_x Coordonnée x à convertir
 #' @param coord_y Coordonnée y à convertir
 #' @keywords SIG
@@ -9,9 +9,9 @@
 #' @import tidyverse
 #' @export
 #' @examples
-#' SIG.conversion.l93dec2(888227, 6655627) %>% mapview::mapview()
+#' SIG.conversion.l932dec(888227, 6655627) %>% mapview::mapview()
 
-    SIG.conversion.l93dec2 <- function(coord_x = NA, coord_y = NA) {
+    SIG.conversion.l932dec <- function(coord_x = NA, coord_y = NA) {
         
         #### Test de cohérence ####
         if(is.na(coord_x)) stop("Pas de coord_x fournie")